カスタムフックselectedPref, ShowSelector, ViewChart
