+++ app/hooks/useGetAllPrefData.ts
@@ -1,5 +1,4 @@
 import { useEffect, useState } from "react";
-import { getPrefData } from "../components/getData/getData";
 
 // 全都道府県のデータを取得するカスタムフック
 const useGetAllPrefData = () => {
@@ -10,6 +9,28 @@ const useGetAllPrefData = () => {
     // 全都道府県のデータを管理
     const [allPrefData, setAllPrefData] = useState<prefDataModel[]|null>(null);
 
+    // 全都道府県のデータを取得する関数
+    const getPrefData = async () => {
+
+        const res = await fetch('api/resas/getPref', {
+        method: 'GET',
+        headers: {
+            'Content-Type': 'application/json; charset=UTF-8',
+
+            }
+        })
+        
+        const result = await res.json();
+        const data_final: prefDataModel[] = result.data.result
+        
+        if (!data_final) {
+            return null;
+        }
+        
+        console.log("都道府県: ", data_final)
+        return data_final;
+    }
+
     useEffect(() => {
         setLoading(false);
 
@@ -18,11 +39,11 @@ const useGetAllPrefData = () => {
         }
         
         if (!allPrefData) {
-        getData();
+            getData();
         }
 
         setLoading(true);
-    },[])
+    }, [])
 
     return { allPrefData, setAllPrefData, loading }
 }
